+++ src/javascript/007_constants.js
@@ -1,5 +1,5 @@
-s.lastEdited = "2019-01-21-0739";
-s.versionNumber = 106;
+s.lastEdited = "2019-01-22-0813";
+s.versionNumber = 107;
 
 s.calcVersion = function(num) {
   let main = Math.floor(num/100);